# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Ámbar Core API
  version: 0.1.0
  summary: API del core bancario Ámbar (ledger, transferencias, dispositivos, avisos y fraude).
  description: |
    Contrato fuente de verdad del core bancario Ámbar. Todo el código cliente
    (apps móvil y web) se genera desde este archivo y QA lo usa para el mock server.

    Convenciones:
    - Montos en **centavos** como enteros (`125050` = $1,250.50 MXN). Nunca decimales.
    - Toda operación que mueve dinero exige el header `Idempotency-Key` (UUID).
      Repetir la misma llave con el mismo cuerpo devuelve la respuesta original
      con `Idempotent-Replayed: true`; con un cuerpo distinto devuelve 409.
    - Errores en formato RFC 9457 (`application/problem+json`) con un `code` estable.
    - Un recurso de otro usuario responde 404, no 403, para no revelar que existe.
  contact:
    name: José Carlos Bernal Bautista
    email: josecarlosbernal09@gmail.com
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3000
    description: Local
  - url: https://api.staging.ambar.example
    description: Staging (datos sintéticos)
...
security:
  - cognito: []
paths:
...
  /v1/accounts:
    get:
      tags: [Cuentas]
      operationId: listAccounts
      summary: Lista las cuentas del usuario autenticado
      description: Devuelve solo las cuentas cuyo dueño es el `sub` del token.
      security:
        - cognito: [accounts.read]
      responses:
        '200':
          description: Cuentas del usuario
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Cuentas]
      operationId: openAccount
      summary: Abre una cuenta de débito en MXN para el usuario autenticado
      description: Asigna una CLABE nueva con dígito verificador válido y publica `account.opened` en el outbox.
      security:
        - cognito: [accounts.write]
      responses:
        '201':
          description: Cuenta abierta con CLABE asignada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/accounts/{accountId}:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags: [Cuentas]
      operationId: getAccount
      summary: Consulta una cuenta y su saldo
      description: Si la cuenta pertenece a otro usuario responde 404 (protección BOLA).
      security:
        - cognito: [accounts.read]
      responses:
        '200':
          description: Cuenta
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/accounts/{accountId}/movements:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags: [Cuentas]
      operationId: listMovements
      summary: Movimientos de la cuenta, del más reciente al más antiguo
      description: Paginación por cursor. El monto se expresa desde la perspectiva del cliente.
      security:
        - cognito: [accounts.read]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: cursor
          in: query
          description: Valor `next_cursor` de la página anterior.
          schema: { type: string, pattern: '^[0-9]+$' }
      responses:
        '200':
          description: Página de movimientos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MovementPage' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/transfers:
    post:
      tags: [Transferencias]
      operationId: createTransfer
      summary: Transfiere dinero a otra cuenta Ámbar identificada por CLABE
      description: |
        Fase 1 solo admite destinos dentro de Ámbar. La transferencia se registra
        en un único asiento de doble partida y en la misma transacción se escribe
        el evento `transfer.posted` en la tabla outbox.

        **Step-up (app móvil).** Si el monto alcanza el umbral o el destino es nuevo
        para el usuario, responde 401 con `code: STEP_UP_REQUIRED` y
        `WWW-Authenticate: Bearer error="insufficient_user_authentication"` (RFC 9470).
        La app pide un reto en `POST /v1/step-up/challenges`, lo firma con la llave
        biométrica del dispositivo y repite la transferencia con el header `X-Step-Up`
        y la misma `Idempotency-Key`.
      security:
        - cognito: [transfers.write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/StepUpProof'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransferRequest' }
            examples:
              renta:
                summary: Pago de renta a otra cuenta Ámbar
                value:
                  source_account_id: 0b6c2f0e-6c47-4f5e-9b0e-2b9a4d1f7c11
                  destination_clabe: '999180000000000015'
                  amount: 650000
                  concept: Renta septiembre
      responses:
        '201':
          description: Transferencia registrada (o repetición idempotente)
          headers:
            Idempotent-Replayed:
              description: '`true` cuando la respuesta es la original de una solicitud repetida.'
              schema: { type: string, enum: ['true', 'false'] }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transfer' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/UnauthenticatedOrStepUp' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '422': { $ref: '#/components/responses/BusinessRule' }
  /v1/devices:
    get:
      tags: [Dispositivos]
      operationId: listDevices
      summary: Dispositivos registrados del usuario
      description: Solo los del usuario autenticado; nunca expone la llave pública completa.
      security:
        - cognito: [devices.write]
      responses:
        '200':
          description: Dispositivos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Dispositivos]
      operationId: registerDevice
      summary: Registra la llave pública de hardware de un dispositivo
      description: |
        La app genera un par ECDSA P-256 en Android Keystore o Secure Enclave, protegido
        con biometría, y envía solo la llave pública (SPKI DER en base64). La llave
        privada nunca sale del hardware. Máximo 3 dispositivos activos por usuario.
      security:
        - cognito: [devices.write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceRegistration' }
      responses:
        '201':
          description: Dispositivo registrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Device' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/BusinessRule' }
  /v1/devices/{deviceId}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    delete:
      tags: [Dispositivos]
      operationId: revokeDevice
      summary: Revoca un dispositivo (por ejemplo, si se perdió el teléfono)
      description: Sus retos pendientes dejan de ser válidos de inmediato.
      security:
        - cognito: [devices.write]
      responses:
        '204':
          description: Dispositivo revocado
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/step-up/challenges:
    post:
      tags: [Dispositivos]
      operationId: createStepUpChallenge
      summary: Crea un reto de un solo uso ligado a una operación concreta
      description: |
        El reto dura 60 segundos y solo sirve para la operación descrita: el
        `signing_payload` incluye el hash de la transferencia. Una firma para
        $100 a una CLABE no sirve para $10,000 a otra.
      security:
        - cognito: [transfers.write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StepUpChallengeRequest' }
      responses:
        '201':
          description: Reto creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StepUpChallenge' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/entries/{entryId}:
    parameters:
      - $ref: '#/components/parameters/EntryId'
    get:
      tags: [Ledger]
      operationId: getEntry
      summary: Consulta un asiento con sus postings (operación)
      description: Vista contable completa del asiento, con signos de débito y crédito. Solo operación.
      security:
        - cognito: [ledger.admin]
      responses:
        '200':
          description: Asiento
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Entry' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/entries/{entryId}/reversals:
    parameters:
      - $ref: '#/components/parameters/EntryId'
    post:
      tags: [Ledger]
      operationId: reverseEntry
      summary: Registra un asiento de reverso (el original nunca se modifica)
      description: Crea un asiento con los postings invertidos. Un asiento solo se reversa una vez y un reverso no se reversa.
      security:
        - cognito: [ledger.admin]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReversalRequest' }
      responses:
        '201':
          description: Asiento de reverso registrado
          headers:
            Idempotent-Replayed:
              schema: { type: string, enum: ['true', 'false'] }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Entry' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '422': { $ref: '#/components/responses/BusinessRule' }
  /v1/qa/deposits:
    post:
      tags: [QA]
      operationId: simulateDeposit
      summary: Simula un depósito SPEI entrante (solo local, dev y staging)
      description: Acredita la cuenta contra la cuenta de liquidación SPEI. El servicio no registra esta ruta en producción.
      x-environments: [local, dev, staging]
      security:
        - cognito: [qa.write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepositRequest' }
      responses:
        '201':
          description: Depósito registrado
          headers:
            Idempotent-Replayed:
              schema: { type: string, enum: ['true', 'false'] }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Entry' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '422': { $ref: '#/components/responses/BusinessRule' }
  /v1/devices/{deviceId}/push-token:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    put:
      tags: [Dispositivos]
      operationId: setDevicePushToken
      summary: Registra o borra el token de notificaciones push del dispositivo
      description: Se borra automáticamente al revocar el dispositivo. `null` lo elimina.
      security:
        - cognito: [devices.write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PushTokenRequest' }
      responses:
        '204':
          description: Token actualizado
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/notifications:
    get:
      tags: [Avisos]
      operationId: listNotifications
      summary: Últimos 50 avisos del usuario
      description: Incluye el número de avisos sin leer para el indicador de la app.
      security:
        - cognito: [accounts.read]
      responses:
        '200':
          description: Avisos, del más reciente al más antiguo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationInbox' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/notifications/{notificationId}/read:
    parameters:
      - $ref: '#/components/parameters/NotificationId'
    post:
      tags: [Avisos]
      operationId: markNotificationRead
      summary: Marca un aviso como leído
      description: Idempotente; conserva la fecha de la primera lectura.
      security:
        - cognito: [accounts.write]
      responses:
        '204':
          description: Aviso marcado como leído
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/fraud-cases/{caseId}:
    parameters:
      - $ref: '#/components/parameters/FraudCaseId'
    get:
      tags: [Avisos]
      operationId: getFraudCase
      summary: Consulta un caso de posible fraude sobre una transferencia propia
      description: Solo los casos del usuario autenticado (404 en otro caso).
      security:
        - cognito: [accounts.read]
      responses:
        '200':
          description: Caso
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FraudCase' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/fraud-cases/{caseId}/answer:
    parameters:
      - $ref: '#/components/parameters/FraudCaseId'
    post:
      tags: [Avisos]
      operationId: answerFraudCase
      summary: El cliente indica si reconoce la transferencia
      description: |
        Si responde que **no** la reconoce, la cuenta de origen se congela en la misma
        transacción (puede recibir dinero, no enviarlo) y se emite `fraud_case.answered`
        para que operación dé seguimiento. Repetir la misma respuesta es idempotente;
        una respuesta distinta a un caso cerrado devuelve 409.
      security:
        - cognito: [transfers.write]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FraudCaseAnswer' }
      responses:
        '200':
          description: Caso resuelto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FraudCaseResolution' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/admin/accounts/{accountId}/unfreeze:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    post:
      tags: [Ledger]
      operationId: unfreezeAccount
      summary: Descongela una cuenta después de revisar el caso (operación)
      description: Idempotente. Emite `account.unfrozen` con el operador y el motivo.
      security:
        - cognito: [ledger.admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UnfreezeRequest' }
      responses:
        '200':
          description: Cuenta activa
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/internal/notifications:
    post:
      tags: [Interna]
      operationId: createNotification
      summary: Crea un aviso para un cliente (n8n)
      description: |
        Idempotente por (`event_id`, destinatario, `kind`): si el flujo se reintenta, devuelve
        el aviso original con 200 y no se duplica. La respuesta incluye los tokens push de los
        dispositivos activos del cliente para que n8n envíe la notificación.
      x-internal: true
      security:
        - cognitoM2M: [internal.notify]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InternalNotificationRequest' }
      responses:
        '201':
          description: Aviso creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalNotificationResult' }
        '200':
          description: El aviso ya existía
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalNotificationResult' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/internal/fraud-cases:
    post:
      tags: [Interna]
      operationId: openFraudCase
      summary: Abre un caso de posible fraude sobre una transferencia (n8n)
      description: |
        n8n decide *cuándo* (puntaje del evento `transfer.posted`); los datos del caso salen
        del asiento, no de la solicitud. Crea en la misma transacción el aviso `FRAUD_CHECK`
        con un texto fijo del core. Idempotente por `entry_id`.
      x-internal: true
      security:
        - cognitoM2M: [internal.fraud]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InternalFraudCaseRequest' }
      responses:
        '201':
          description: Caso abierto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalFraudCaseResult' }
        '200':
          description: El caso ya existía
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalFraudCaseResult' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/internal/reconciliation:
    get:
      tags: [Interna]
      operationId: getReconciliation
      summary: Conciliación del ledger y salud del outbox (n8n, diaria)
      description: |
        `ok` es falso si algún saldo no coincide con sus postings, si el ledger no suma cero,
        si hay eventos muertos o si el evento pendiente más antiguo tiene más de 5 minutos.
      x-internal: true
      security:
        - cognitoM2M: [internal.reconcile]
      responses:
        '200':
          description: Reporte
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReconciliationReport' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    cognito:
      type: oauth2
      description: |
        Amazon Cognito (Authorization Code + PKCE). Los scopes se emiten con el
        prefijo del resource server, p. ej. `ambar-api/transfers.write`.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.ambar.example/oauth2/authorize
          tokenUrl: https://auth.ambar.example/oauth2/token
          refreshUrl: https://auth.ambar.example/oauth2/token
          scopes:
            accounts.read: Consultar cuentas y movimientos propios
            accounts.write: Abrir cuentas
            transfers.write: Enviar transferencias
            devices.write: Registrar y revocar dispositivos de la app móvil
            ledger.admin: Operación del ledger (consultar y reversar asientos)
            qa.write: Simular movimientos en entornos no productivos
            assistant.chat: Conversar con el asistente financiero
    cognitoM2M:
      type: oauth2
      description: Cognito client_credentials para la automatización (n8n). Solo dentro de la VPC.
      flows:
        clientCredentials:
          tokenUrl: https://auth.ambar.example/oauth2/token
          scopes:
            internal.notify: Crear avisos para clientes
            internal.fraud: Abrir casos de posible fraude
            internal.reconcile: Consultar la conciliación del ledger y del outbox
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    StepUpProof:
      name: X-Step-Up
      in: header
      required: false
      description: '`<challenge_id>.<firma>`: la firma ECDSA (DER, base64url) del `signing_payload` del reto.'
      schema: { type: string, pattern: '^[0-9a-f-]{36}\.[A-Za-z0-9_-]{16,200}$' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID generado por el cliente; se reutiliza en cada reintento de la misma operación.
      schema: { type: string, format: uuid }
  schemas:
    Uuid:
      type: string
      format: uuid
...
    Account:
      type: object
      additionalProperties: false
      required: [id, clabe, currency, status, balance, created_at]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        clabe: { $ref: '#/components/schemas/Clabe' }
        currency: { $ref: '#/components/schemas/Currency' }
        status:
          type: string
          enum: [ACTIVE, FROZEN, CLOSED]
        balance:
          type: integer
          format: int64
          minimum: 0
          description: Saldo disponible en centavos.
        created_at: { $ref: '#/components/schemas/Timestamp' }
    AccountList:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Account' }
...
    MovementPage:
      type: object
      additionalProperties: false
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Movement' }
        next_cursor:
          type: [string, 'null']
    TransferRequest:
      type: object
      additionalProperties: false
      required: [source_account_id, destination_clabe, amount, concept]
      properties:
        source_account_id: { $ref: '#/components/schemas/Uuid' }
        destination_clabe: { $ref: '#/components/schemas/Clabe' }
        amount: { $ref: '#/components/schemas/Amount' }
        concept:
          type: string
          minLength: 1
          maxLength: 40
          description: Concepto de pago (máximo 40 caracteres, como en SPEI).
...
  responses:
    ValidationError:
      description: La solicitud no cumple el contrato
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthenticated:
      description: Falta el token o es inválido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Forbidden:
      description: El token no tiene el scope requerido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: El recurso no existe o no pertenece al usuario
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...